fix FoodStoreResponse has double but response has float, standardize - v0.13.1-f13

+++ recommendation-service/src/main/java/com/gokdenizozkan/ddd/recommendationservice/core/quality/Recommendable.java
@@ -1,4 +1,4 @@
 package com.gokdenizozkan.ddd.recommendationservice.core.quality;
 
-public interface Recommendable<ID> extends Appraisable<Double>, IdHolder<ID>, NameHolder, RatingHolder<Double> {
+public interface Recommendable<ID> extends Appraisable<Float>, IdHolder<ID>, NameHolder, RatingHolder<Float> {
 }
